requirements.txt
aiomysql==0.2.0
aiosqlite==0.20.0
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.4.1
asyncmy==0.2.9
bcrypt==4.1.3
certifi==2024.7.4
cffi==1.16.0
click==8.0.1
colorama==0.4.4
cryptography==42.0.8
dnspython==2.6.1
email_validator==2.2.0
fastapi==0.111.0
fastapi-cli==0.0.4
fastapi-jwt-auth==0.5.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
orjson==3.10.6
passlib==1.7.4
pycparser==2.22
pydantic==2.7.4
pydantic_core==2.18.4
Pygments==2.18.0
PyJWT==1.7.1
PyMySQL==1.1.1
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
rich==13.7.1
setuptools==70.3.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.27
SQLAlchemy-Utils==0.37.8
starlette==0.37.2
typer==0.12.3
typing_extensions==4.8.0
ujson==5.10.0
uvicorn==0.30.1
watchfiles==0.22.0
websockets==12.0

